requirements.txt
autopep8==1.5.6
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
googlesearch-python==1.1.0
idna==2.10
joblib==1.0.1
numpy==1.20.1
pycodestyle==2.7.0
python-dotenv==0.15.0
requests==2.25.1
scikit-learn==0.24.1
scipy==1.6.1
sklearn==0.0
soupsieve==2.0.1
threadpoolctl==2.1.0
toml==0.10.2
urllib3==1.25.9
whois==0.9.8